Fix compilation errors

+++ blocktx/store/sql/get_block_transactions_test.go
@@ -27,7 +27,7 @@ func (s *GetBlockTransactionsSuite) Run() {
 
 	s.InsertBlockTransactionMap(&store.BlockTransactionMap{
 		BlockID:       block.ID,
-		TransactionID: tx.ID,
+		TransactionID: int64(tx.ID),
 		Pos:           2,
 	})
